requirements.txt
Django>=4.2.0
PyYAML>=6.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
simplejson>=3.19.0
python-dateutil>=2.8.2
djangorestframework-simplejwt>=5.2.0
django-filter>=22.1
pyyaml>=6.0.0
openapi-codec>=1.3.2
